Finished coding invoices, need data to test so commiting this to test with main

+++ Ver_0.01/Team34_GP_IFM02B2_2023_WebApp/Team34_GP_IFM02B2_2023_WebApp/Connected Services/ResQReference/Reference.cs
@@ -1580,10 +1580,10 @@ namespace Team34_GP_IFM02B2_2023_WebApp.ResQReference {
         System.Threading.Tasks.Task<bool> editStoreAsync(Team34_GP_IFM02B2_2023_WebApp.ResQReference.Store S);
         
         [System.ServiceModel.OperationContractAttribute(Action="http://tempuri.org/IRESQSERVICE/addInvoice", ReplyAction="http://tempuri.org/IRESQSERVICE/addInvoiceResponse")]
-        bool addInvoice(int UID, double price, System.DateTime TOS, Team34_GP_IFM02B2_2023_WebApp.ResQReference.CartItem[] cart);
+        bool addInvoice(int UID, decimal price, System.DateTime TOS, Team34_GP_IFM02B2_2023_WebApp.ResQReference.CartItem[] cart);
         
         [System.ServiceModel.OperationContractAttribute(Action="http://tempuri.org/IRESQSERVICE/addInvoice", ReplyAction="http://tempuri.org/IRESQSERVICE/addInvoiceResponse")]
-        System.Threading.Tasks.Task<bool> addInvoiceAsync(int UID, double price, System.DateTime TOS, Team34_GP_IFM02B2_2023_WebApp.ResQReference.CartItem[] cart);
+        System.Threading.Tasks.Task<bool> addInvoiceAsync(int UID, decimal price, System.DateTime TOS, Team34_GP_IFM02B2_2023_WebApp.ResQReference.CartItem[] cart);
         
         [System.ServiceModel.OperationContractAttribute(Action="http://tempuri.org/IRESQSERVICE/addInvoiceItem", ReplyAction="http://tempuri.org/IRESQSERVICE/addInvoiceItemResponse")]
         bool addInvoiceItem(int ID, Team34_GP_IFM02B2_2023_WebApp.ResQReference.CartItem c);
@@ -1803,11 +1803,11 @@ namespace Team34_GP_IFM02B2_2023_WebApp.ResQReference {
             return base.Channel.editStoreAsync(S);
         }
         
-        public bool addInvoice(int UID, double price, System.DateTime TOS, Team34_GP_IFM02B2_2023_WebApp.ResQReference.CartItem[] cart) {
+        public bool addInvoice(int UID, decimal price, System.DateTime TOS, Team34_GP_IFM02B2_2023_WebApp.ResQReference.CartItem[] cart) {
             return base.Channel.addInvoice(UID, price, TOS, cart);
         }
         
-        public System.Threading.Tasks.Task<bool> addInvoiceAsync(int UID, double price, System.DateTime TOS, Team34_GP_IFM02B2_2023_WebApp.ResQReference.CartItem[] cart) {
+        public System.Threading.Tasks.Task<bool> addInvoiceAsync(int UID, decimal price, System.DateTime TOS, Team34_GP_IFM02B2_2023_WebApp.ResQReference.CartItem[] cart) {
             return base.Channel.addInvoiceAsync(UID, price, TOS, cart);
         }
